requirements.txt
python-telegram-bot==20.8
psycopg[binary]
python-dotenv
flask
requests
phonenumbers